+++ __fixtures__/misc/output-decimals/agoric/gogoproto/bundle.ts
@@ -1,4 +1,0 @@
-import * as _170 from "./gogo";
-export const gogoproto = {
-  ..._170
-};