+++ __fixtures__/misc/output-decimals/gogoproto/bundle.ts
@@ -1,0 +1,4 @@
+import * as _182 from "./gogo";
+export const gogoproto = {
+  ..._182
+};