+++ __fixtures__/misc/output-decimals/google/bundle.ts
@@ -1,0 +1,20 @@
+import * as _183 from "./api/expr/v1alpha1/eval";
+import * as _184 from "./protobuf/any";
+import * as _185 from "./protobuf/descriptor";
+import * as _186 from "./protobuf/duration";
+import * as _187 from "./protobuf/timestamp";
+export namespace google {
+  export namespace api {
+    export namespace expr {
+      export const v1alpha1 = {
+        ..._183
+      };
+    }
+  }
+  export const protobuf = {
+    ..._184,
+    ..._185,
+    ..._186,
+    ..._187
+  };
+}