+++ __fixtures__/misc/output-proto-amino/amino-msg-no-func/google/bundle.ts
@@ -1,20 +1,20 @@
-import * as _115 from "./api/expr/v1alpha1/eval";
-import * as _116 from "./protobuf/any";
-import * as _117 from "./protobuf/descriptor";
-import * as _118 from "./protobuf/duration";
-import * as _119 from "./protobuf/timestamp";
+import * as _127 from "./api/expr/v1alpha1/eval";
+import * as _128 from "./protobuf/any";
+import * as _129 from "./protobuf/descriptor";
+import * as _130 from "./protobuf/duration";
+import * as _131 from "./protobuf/timestamp";
 export namespace google {
   export namespace api {
     export namespace expr {
       export const v1alpha1 = {
-        ..._115
+        ..._127
       };
     }
   }
   export const protobuf = {
-    ..._116,
-    ..._117,
-    ..._118,
-    ..._119
+    ..._128,
+    ..._129,
+    ..._130,
+    ..._131
   };
 }