+++ __fixtures__/misc/output-proto-amino/proto-only-legacy/gogoproto/bundle.ts
@@ -1,4 +1,4 @@
-import * as _88 from "./gogo";
+import * as _100 from "./gogo";
 export const gogoproto = {
-  ..._88
+  ..._100
 };