+++ packages/ast/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.5](https://github.com/cosmology-tech/telescope/compare/@cosmology/ast@1.8.4...@cosmology/ast@1.8.5) (2024-10-05)
+
+**Note:** Version bump only for package @cosmology/ast
+
 ## [1.8.4](https://github.com/cosmology-tech/telescope/compare/@cosmology/ast@1.8.3...@cosmology/ast@1.8.4) (2024-09-18)
 
 **Note:** Version bump only for package @cosmology/ast