+++ packages/starship/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.7](https://github.com/osmosis-labs/telescope/compare/@cosmology/starship-test@1.8.6...@cosmology/starship-test@1.8.7) (2024-10-05)
+
+**Note:** Version bump only for package @cosmology/starship-test
+
 ## [1.8.6](https://github.com/osmosis-labs/telescope/compare/@cosmology/starship-test@1.8.5...@cosmology/starship-test@1.8.6) (2024-09-18)
 
 **Note:** Version bump only for package @cosmology/starship-test