+++ packages/types/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.1](https://github.com/cosmology-tech/telescope/compare/@cosmology/types@1.10.0...@cosmology/types@1.10.1) (2024-10-05)
+
+**Note:** Version bump only for package @cosmology/types
+
 # [1.10.0](https://github.com/cosmology-tech/telescope/compare/@cosmology/types@1.9.0...@cosmology/types@1.10.0) (2024-09-18)
 
 ### Features